chore(project): Initialized dev environment and tidied main entry

+++ src-server/app/main.c
@@ -11,18 +11,16 @@
 /*
  * TODO List:
  *  - Implement /src-server/infra/fs
+ *  - Study networking to start implementing /src-server/infra/socket
  *
  */
 
 void init() {
-    Random.seed(time(NULL));
-    Random.seed(Random.rand());
-
+    Random.seed((int)time(NULL));
+    Random.seed((int)Random.rand());
 }
 
 int main() {
     init();
     return 0;
-    
 }
-
